+++ package.json
@@ -69,7 +69,7 @@
     "ci-info": "^4.4.0",
     "local-pkg": "^1.2.1",
     "selfsigned": "^5.5.0",
-    "vite-plugin-electron": "~1.0.0"
+    "vite-plugin-electron": "~1.0.1"
   },
   "devDependencies": {
     "@subf/config": "^0.2.1",
