+++ src/vite/define.ts
@@ -1,7 +1,7 @@
 import type { InlineConfig, UserConfig, UserConfigFn } from 'vite'
 
 import { electronWithUpdater } from './core'
-import type { ElectronWithUpdaterOptions } from './option'
+import type { ElectronWithUpdaterOptions } from './types'
 
 type MakeOptional<T, K extends keyof T> = Partial<Pick<T, K>> & Omit<T, K>
