+++ src/vite/index.ts
@@ -3,4 +3,4 @@ export { electronWithUpdater } from './core'
 export { electronWithUpdater as default } from './core'
 export { defineElectronConfig } from './define'
 export type { ElectronViteHelperOptions } from './define'
-export type { ElectronWithUpdaterOptions } from './option'
+export type { ElectronWithUpdaterOptions } from './types'
