+++ src/vite/type.d.ts
@@ -1,5 +1,6 @@
 declare namespace NodeJS {
   interface Process {
     electronApp: import('node:child_process').ChildProcess
+    cachedOptions: import('./types').ResolvedOptions
   }
 }
